refactor to client, types, worker file
