+++ src/logic/FindAllReferences.ts
@@ -1,9 +1,10 @@
 import { BehaviorSubject, combineLatest, distinctUntilChanged, from, map, Observable, switchMap, throttleTime } from "rxjs";
-import { jarIndex, type ReferenceKey, type ReferenceString } from "../workers/JarIndex";
+import { jarIndex } from "../workers/jar-index/client";
 import { openCodeTab } from "./Tabs";
 import { referencesQuery } from "./State";
 import type { Token } from "./Tokens";
 import type { DecompileResult } from "../workers/decompile/types";
+import type { ReferenceKey, ReferenceString } from "../workers/jar-index/types";
 
 export const referenceResults = referencesQuery
     .pipe(
