--- src/workers/JarIndexWorker2.ts
@@ -1,4 +1,0 @@
-import * as Comlink from "comlink";
-import { JarIndexWorker } from "./JarIndexWorker";
-
-Comlink.expose(new JarIndexWorker());
