+++ src/workers/jar-index/types.ts
@@ -1,11 +1,22 @@
-import { load } from "../../java/build/generated/teavm/wasm-gc/java.wasm-runtime.js";
-import indexerWasm from '../../java/build/generated/teavm/wasm-gc/java.wasm?url';
-import { openJar, type Jar } from "../utils/Jar.js";
-import type { ReferenceKey, ReferenceString } from "./JarIndex.js";
+import { load } from "../../../java/build/generated/teavm/wasm-gc/java.wasm-runtime.js";
+import indexerWasm from '../../../java/build/generated/teavm/wasm-gc/java.wasm?url';
+import { openJar, type Jar } from "../../utils/Jar.js";
+
+export type Class = string;
+export type Method = `${string}:${string}:${string}`;
+export type Field = `${string}:${string}:${string}`;
+
+// oxlint-disable-next-line typescript/no-redundant-type-constituents
+export type ReferenceKey = Class | Method;
+
+export type ReferenceString =
+    | `c:${Class}`
+    | `m:${Method}`
+    | `f:${Field}`;
 
 export type ClassDataString = `${string}|${string}|${number}|${string}`;
 
-export class JarIndexWorker {
+export class JarIndexer {
     #indexerFunc: Indexer | null = null;
     #jar: Jar | null = null;
 
@@ -16,13 +27,13 @@ export class JarIndexWorker {
                 this.#indexerFunc = teavm.exports as Indexer;
             } catch (e) {
                 console.warn("Failed to load WASM module (non-compliant browser?), falling back to JS implementation", e);
-                this.#indexerFunc = await import("../../java/build/generated/teavm/js/java.js") as unknown as Indexer;
+                this.#indexerFunc = await import("../../../java/build/generated/teavm/js/java.js") as unknown as Indexer;
             }
         }
         return this.#indexerFunc;
     };
 
-    setWorkerJar = async (name: string, blob: Blob | null) => {
+    setJar = async (name: string, blob: Blob | null) => {
         if (!blob) {
             this.#jar = null;
             return;
