+++ src/workers/jar-index/worker.ts
@@ -1,0 +1,4 @@
+import * as Comlink from "comlink";
+import { JarIndexer } from "./types";
+
+Comlink.expose(new JarIndexer());
